% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.wcov}
\alias{.wcov}
\title{Weighted Covariance Matrix}
\usage{
.wcov(x, w)
}
\arguments{
\item{x}{A matrix with \eqn{m} columns and \eqn{n} rows, where each column 
represents a different variable and each row a different observation.}

\item{w}{A non-negative and non-zero vector of weights for each observation. 
Its length must equal the number of rows of x.}
}
\value{
A list with two components:
 \itemize{
  \item \code{center}: an estimate for the center (mean) of the data.
  \item \code{cov}: the estimated (weighted) covariance matrix.
 }
}
\description{
\code{wcov} computes the estimates of the weighted covariance 
 matrix and the weighted mean of the data.
}
\examples{
m <- matrix(c(rnorm(500, 6), rnorm(500, 11, 3)), ncol = 2)
w <- runif(500)
boRing:::.wcov(m, w)

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
