% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boRing-package.R
\docType{data}
\name{four_Gaussians}
\alias{four_Gaussians}
\title{Four Gaussian Clusters}
\description{
Matrix of 2-dimensional points forming four Gaussian clusters.
}
\examples{
data(four_Gaussians)
plot(four_Gaussians, pch = 19)

}
\keyword{datasets}
