% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mahalanobis}
\alias{Mahalanobis}
\title{Mahalanobis Distance}
\usage{
Mahalanobis(x, center, cov)
}
\arguments{
\item{x}{A matrix with \eqn{m} columns and \eqn{n} rows, where each column 
represents a different variable and each row a different observation.}

\item{center}{The mean vector of the distribution.}

\item{cov}{The covariance matrix (\eqn{p \times p}{p x p}) of the 
distribution.}
}
\value{
A matrix with 1 column and \eqn{n} rows reprensenting the estimated
 distance for each observation.
}
\description{
\code{Mahalanobis} computes the squared Mahalanobis distance of
 all rows in \code{x} and the vector \eqn{\mu}{mu} = \code{center} with 
 respect to \eqn{\Sigma}{Sigma} = \code{cov}. This is (for vector \code{x}) 
 defined as \deqn{D^2 = (x - \mu)' \Sigma^{-1} (x - \mu)}{D^2 = (x - \mu)' 
 \Sigma^-1 (x - \mu)}
}
\examples{
m <- matrix(c(rnorm(500, 6), rnorm(500, 11, 3)), ncol = 2)
w <- runif(500)
covar <- wcov(m, w)
Mahalanobis(m, covar$center, covar$cov)

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
