% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boRing-package.R
\docType{data}
\name{three_Gaussians}
\alias{three_Gaussians}
\title{Three Gaussian Clusters}
\description{
Matrix of 2-dimensional points forming three Gaussian clusters.
}
\examples{
data(three_Gaussians)
plot(three_Gaussians, pch = 19)

}
\keyword{datasets}
